requirements.txt
pika==1.3.2
fastapi==0.115.5
uvicorn==0.32.0
Jinja2==3.1.4
websockets==14.1
faststream==0.5.30
SQLAlchemy==2.0.36
PyJWT==2.10.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
alembic==1.14.0
asyncpg==0.30.0
psycopg2==2.9.10